// src/lib/types/ALObjectType.ts
export enum ALObjectType {
  Table = "table",
  TableExtension = "tableextension",
  Page = "page",
  PageExtension = "pageextension",
  Report = "report",
  ReportExtension = "reportextension",
  Codeunit = "codeunit",
  Query = "query",
  XmlPort = "xmlport",
  Enum = "enum",
  EnumExtension = "enumextension",
  ControlAddIn = "controladdin",
  Interface = "interface",
  PermissionSet = "permissionset",
  PermissionSetExtension = "permissionsetextension"
}